requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
fritzconnection>=1.14
apscheduler>=3.10,<4
beautifulsoup4>=4.12
pydantic-settings>=2.2
